Refactor: Large refactor of the codebase
1. New sub-package aleph_message.models.execution with everything related to function and instance execution.
2. Review of the type annotations, now MyPy passes without any error.
3. Improvements in code coverage.
4. More docstricts and field descriptions
