+++ aleph_message/__init__.py
@@ -1,1 +1,3 @@
-from .models import Message, MessagesResponse
+from .models import MessagesResponse, parse_message
+
+__all__ = ["parse_message", "MessagesResponse"]
