+++ aleph_message/tests/test_status.py
@@ -1,0 +1,5 @@
+from aleph_message.status import MessageStatus
+
+
+def test_message_status():
+    assert MessageStatus.PENDING == "pending"
