+++ aleph_message/tests/test_utils.py
@@ -1,6 +1,6 @@
-from aleph_message.utils import gigabyte_to_mebibyte
+from aleph_message.utils import Gigabytes, gigabyte_to_mebibyte
 
 
 def test_gigabyte_to_mebibyte():
-    assert gigabyte_to_mebibyte(1) == 954
-    assert gigabyte_to_mebibyte(100) == 95368
+    assert gigabyte_to_mebibyte(Gigabytes(1)) == 954
+    assert gigabyte_to_mebibyte(Gigabytes(100)) == 95368
