+++ aleph_message/utils.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import math
 from typing import NewType
